# apps/api/docs/openapi.yml
openapi: 3.0.4
info:
  title: terraforge
  version: 0.0.1

servers:
  - url: https://staging-api.terraforge.gg/v1
    description: Staging server
  - url: https://api.terraforge.gg/v1
    description: Production server

security:
  - bearerAuth: []

paths:
  /projects:
    get:
      summary: Search projects
      parameters:
        - name: query
          schema:
            type: string
          in: query
          required: false
        - name: type
          schema:
            type: string
          in: query
          required: false
        - name: limit
          schema:
            type: integer
            format: int64
            minimum: 0
          in: query
          required: false
        - name: offset
          schema:
            type: integer
            format: int64
            minimum: 0
            maximum: 100
          in: query
          required: false
      tags:
        - Projects
      responses:
        "201":
          description: Project created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectSearch"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
    post:
      summary: Create a new project
      tags:
        - Projects
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectRequest"
      responses:
        "201":
          description: Project created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
  /projects/{id|slug}:
    parameters:
      - $ref: "#/components/parameters/ProjectIdentifier"
    get:
      tags:
        - Projects
      summary: Find project by id or slug
      security: []
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
    patch:
      tags:
        - Projects
      summary: Update project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectRequest"
      responses:
        "204":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
    delete:
      tags:
        - Projects
      summary: Delete project
      responses:
        "204":
          description: successful operation
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
  /projects/{id|slug}/members:
    parameters:
      - $ref: "#/components/parameters/ProjectIdentifier"
    get:
      tags:
        - Projects
      summary: List project members
      security: []
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                minItems: 0
                items:
                  $ref: "#/components/schemas/ProjectMember"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
  /loader-versions:
    get:
      tags:
        - Loader versions
      summary: List of loader versions
      security: []
      responses:
        '200':
          description: successful operation
          content:
              application/json:
                schema:
                  type: array
                  minItems: 0
                  items:
                    $ref: '#/components/schemas/LoaderVersion'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /projects/{id|slug}/releases/upload-url:
    parameters:
      - $ref: '#/components/parameters/ProjectIdentifier'
    get:
      tags:
        - Projects
      summary: Generate a presigned put url
      parameters:
        - name: fileSize
          in: query
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
              application/json:
                schema:
                  type: string
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '401':
          description: Unauthorised
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /projects/{id|slug}/releases:
    parameters:
      - $ref: '#/components/parameters/ProjectIdentifier'
    get:
      tags:
        - Projects
      summary: Get project releases
      responses:
        '200':
          description: successful operation
          content:
              application/json:
                schema:
                  type: array
                  minItems: 0
                  items:
                    $ref: '#/components/schemas/ProjectRelease'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
    post:
      tags:
        - Projects
      summary: Create project release
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectReleaseRequest'
      responses:
        '201':
          description: Project release created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectRelease'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
components:
  parameters:
    ProjectIdentifier:
      name: id|slug
      in: path
      required: true
      description: The id or slug of the project
      schema:
        type: string
  schemas:
    Project:
      type: object
      required:
        - id
        - name
        - slug
        - downloads
        - type
        - status
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          description: Unique identifier of the project
        name:
          type: string
          description: Name of the project
        slug:
          type: string
          description: URL-friendly identifier
        summary:
          type: string
          nullable: true
          description: Short summary of the project
        description:
          type: string
          nullable: true
          description: Detailed description of the project
        iconUrl:
          type: string
          format: uri
          nullable: true
          description: URL to the project's icon
        downloads:
          type: integer
          format: int64
          description: Number of downloads
        type:
          $ref: "#/components/schemas/ProjectType"
        status:
          $ref: "#/components/schemas/ProjectStatus"
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
        userId:
          type: string
          description: Identifier of the user
    ProjectType:
      type: string
      enum:
        - mod
    ProjectStatus:
      type: string
      enum:
        - draft
        - rejected
        - approved
        - banned
    ProjectMember:
      type: object
      required:
        - username
        - userId
        - role
      properties:
        username:
          type: string
        userId:
          type: string
        image:
          type: string
        role:
          $ref: "#/components/schemas/ProjectMemberRole"
    ProjectMemberRole:
      type: string
      enum:
        - owner
        - admin
        - developer
        - maintainer
        - member
    ProjectSearch:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Project"
        limit:
          type: integer
          format: int64
        offset:
          type: integer
          format: int64
        totalHits:
          type: integer
          format: int64
      required:
        - data
        - limit
        - offset
        - totalHits
    LoaderVersionBuildType:
      type: string
      enum:
        - stable
        - preview
        - legacy
    LoaderVersion:
      type: object
      properties:
        id:
          type: string
        gameVersion:
          type: string
        versionLabel:
          type: string
        buildType:
          $ref: '#/components/schemas/LoaderVersionBuildType'
        releasedAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - gameVersion
        - versionLabel
        - buildType
        - releasedAt
        - updatedAt
    ProjectRelease:
      type: object
      properties:
        id:
          type: string
          description: Version id
        projectId:
          type: string
          description: Project id
        name:
          type: string
          description: Version name
        changelog:
          type: string
          description: Version changelog
        versionNumber:
          type: string
        loaderVersion:
          $ref: '#/components/schemas/LoaderVersion'
        downloads:
          type: integer
          format: int64
        fileUrl:
          type: string
        fileSize:
          type: integer
          format: int64
        fileHash:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        publishedAt:
          type: string
          format: date-time
        dependencies:
          type: array
          items:
            $ref: '#/components/schemas/ProjectReleaseDependency'
      required:
        - id
        - projectId
        - name
        - versionNumber
        - loaderVersion
        - downloads
        - fileUrl
        - fileSize
        - fileHash
        - createdAt
        - updatedAt
        - dependencies
    ProjectReleaseDependency:
      type: object
      properties:
        id:
          type: string
          description: id
        releaseId:
          type: string
        dependencyProjectId:
          type: string
        dependencyType:
          $ref: '#/components/schemas/ProjectReleaseDependencyType'
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - releaseId
        - dependencyProjectId
        - dependencyType
        - createdAt
    ProjectReleaseDependencyType:
      type: string
      enum:
        - required
        - optional

    CreateProjectReleaseDependencyRequest:
      type: object
      properties:
        type:
          $ref: '#/components/schemas/ProjectReleaseDependencyType'
        projectId:
          type: string
        minVersionNumber:
          type: string
      required:
        - type
        - projectId
    CreateProjectReleaseRequest:
      type: object
      properties:
        name:
          type: string
        versionNumber:
          type: string
        loaderVersionId:
          type: string
        fileUrl:
          type: string
        dependencies:
          type: array
          items:
            $ref: '#/components/schemas/CreateProjectReleaseDependencyRequest'
      required:
        - name
        - versionNumber
        - loaderVersionId
        - fileUrl
    CreateProjectRequest:
      type: object
      required:
        - type
        - name
        - slug
      properties:
        type:
          $ref: "#/components/schemas/ProjectType"
        name:
          type: string
          minLength: 3
          maxLength: 100
          description: Name of the project
        slug:
          type: string
          minLength: 3
          maxLength: 100
          description: Slug of the project
        summary:
          type: string
          maxLength: 120
          description: Summary of the project
    UpdateProjectRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 3
          maxLength: 100
          description: Name of the project
        slug:
          type: string
          minLength: 3
          maxLength: 100
          description: Slug of the project
        summary:
          type: string
          maxLength: 120
          description: Summary of the project
        description:
          type: string
          description: Description of the project
        iconUrl:
          type: string
          description: Icon url of the project
    ProblemDetails:
      type: object
      required:
        - title
        - status
        - detail
      properties:
        title:
          type: string
          description: A short, human-readable summary of the problem
        status:
          type: integer
          format: int32
          description: The HTTP status code
        detail:
          type: string
          description: A human-readable explanation specific to this occurrence of the problem
        errors:
          type: array
          description: List of validation errors (if applicable)
          items:
            $ref: "#/components/schemas/ValidationError"
    ValidationError:
      type: object
      properties:
        field:
          type: string
          description: Name of the field where the error occurred
        code:
          type: string
          description: Machine-readable error code
        message:
          type: string
          description: Human-readable error message
  
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

tags:
  - name: Projects
    description: Operations related to project management