# apps/api/internal/dto/problem_details.go
package dto

type ProblemDetails struct {
	Title  string            `json:"title"`
	Status int               `json:"status"`
	Detail string            `json:"detail,omitempty"`
	Errors map[string]string `json:"errors,omitempty"`
}